refactor(cli): route native BYOK through runtime adapters
