+++ packages/cli/src/runtime/managed-codex-provider.ts
@@ -1,0 +1,180 @@
+import { mkdirSync, readFileSync } from "node:fs";
+import { join } from "node:path";
+import { MANAGED_AI_PROVIDER_RUNTIME_ENV } from "@clawdi/shared";
+import { writePrivateFileAtomic } from "../lib/private-file";
+import { isValidSemver } from "../lib/semver";
+import type { CatalogProviderConfigurationResult } from "./catalog-provider-config";
+import { normalizeSecretRef } from "./hosted-egress-profiles";
+import type { RuntimeManifest } from "./manifest-contract";
+import { tail } from "./manifest-install";
+import { recordValue, stringValue } from "./manifest-shared";
+import type { RuntimePaths } from "./paths";
+import { runtimeImpactRevision } from "./runtime-impact-revision";
+import { commandExists, executableExists, spawnRuntimeUserCommand } from "./runtime-user-command";
+
+const CODEX_BOOTSTRAP_TIMEOUT_MS = 600_000;
+const CODEX_MANAGED_PROVIDER_ID = "clawdi";
+export const CODEX_MANAGED_PROVIDER_CONFIG_FILE = "config.toml";
+const CODEX_BOOTSTRAP_PACKAGE_VERSION = "0.146.0";
+const CODEX_BOOTSTRAP_PACKAGE_SPEC = `@openai/codex@${CODEX_BOOTSTRAP_PACKAGE_VERSION}`;
+interface HostedCodexManagedProvider {
+	baseUrl: string;
+}
+export function applyHostedCodexManagedProviderProjection(
+	manifest: RuntimeManifest,
+	home: string,
+	codexCli: Record<string, string> | null,
+): CatalogProviderConfigurationResult {
+	const provider = hostedCodexManagedProvider(manifest);
+	if (!provider) return { path: null, revision: null, providerIds: [] };
+
+	const codexHome = hostedCodexHome(home);
+	const configPath = join(codexHome, CODEX_MANAGED_PROVIDER_CONFIG_FILE);
+	const configContent = hostedCodexManagedConfigToml(provider);
+	writePrivateFileAtomic(configPath, configContent, { mode: 0o600, dirMode: 0o700 });
+
+	return {
+		path: configPath,
+		providerIds: [CODEX_MANAGED_PROVIDER_ID],
+		revision: runtimeImpactRevision({
+			codexManagedProviderProjection: CODEX_MANAGED_PROVIDER_CONFIG_FILE,
+			configContent,
+			codexCli,
+		}),
+	};
+}
+export function hostedCodexManagedProvider(
+	manifest: RuntimeManifest,
+): HostedCodexManagedProvider | null {
+	const terminalTooling = recordValue(manifest.projection?.terminalTooling);
+	const codex = recordValue(terminalTooling?.codex);
+	const provider = recordValue(codex?.provider);
+	const primaryModel = recordValue(codex?.primary_model);
+	const providerId = stringValue(codex?.provider_id);
+	const baseUrl = stringValue(provider?.baseUrl);
+	const apiMode = stringValue(provider?.apiMode);
+	const primaryModelName = stringValue(primaryModel?.model);
+	if (
+		codex?.enabled !== true ||
+		!provider ||
+		provider.managed_by !== "clawdi" ||
+		apiMode !== "openai_responses" ||
+		stringValue(provider.runtimeEnvName) !== MANAGED_AI_PROVIDER_RUNTIME_ENV ||
+		normalizeSecretRef(stringValue(provider.apiKeySecretRef)) !== "secret://tool.codex.apiKey" ||
+		!providerId ||
+		stringValue(primaryModel?.provider_id) !== providerId ||
+		!baseUrl ||
+		!primaryModelName
+	) {
+		return null;
+	}
+	return { baseUrl };
+}
+export function hostedCodexHome(home: string): string {
+	return join(home, ".codex");
+}
+export function hostedCodexManagedConfigToml(provider: HostedCodexManagedProvider): string {
+	const lines = ["# Generated by Clawdi hosted runtime. Do not put API keys in this file."];
+	lines.push(
+		`model_provider = ${quoteTomlString(CODEX_MANAGED_PROVIDER_ID)}`,
+		"",
+		`[model_providers.${CODEX_MANAGED_PROVIDER_ID}]`,
+		`name = ${quoteTomlString("clawdi")}`,
+		`base_url = ${quoteTomlString(provider.baseUrl)}`,
+		`env_key = ${quoteTomlString(MANAGED_AI_PROVIDER_RUNTIME_ENV)}`,
+		'wire_api = "responses"',
+		"",
+	);
+	return lines.join("\n");
+}
+export function ensureHostedCodexCli(paths: RuntimePaths): Record<string, string> | null {
+	if (process.env.CLAWDI_CODEX_INSTALL_DISABLED === "1") return null;
+	const npmPrefix = paths.userNpmPrefix;
+	const realBin = join(npmPrefix, "bin", "codex");
+	let installedVersion = hostedCodexInstalledVersion(npmPrefix);
+	const bootstrapRequired = installedVersion === null || !executableExists(realBin);
+	if (bootstrapRequired) {
+		installHostedCodexBootstrap(CODEX_BOOTSTRAP_PACKAGE_SPEC, npmPrefix, paths);
+		installedVersion = hostedCodexInstalledVersion(npmPrefix);
+		if (installedVersion !== CODEX_BOOTSTRAP_PACKAGE_VERSION) {
+			throw new Error(
+				`Codex bootstrap installed version ${installedVersion ?? "unknown"}; expected ${CODEX_BOOTSTRAP_PACKAGE_VERSION}`,
+			);
+		}
+		if (!executableExists(realBin)) {
+			throw new Error(`Codex bootstrap did not create ${realBin}`);
+		}
+	}
+	if (installedVersion === null) throw new Error("Codex package metadata is unavailable");
+	return {
+		commandPath: realBin,
+		npmPrefix,
+		bootstrapPackageSpec: CODEX_BOOTSTRAP_PACKAGE_SPEC,
+		installedVersion,
+		realBin,
+	};
+}
+function hostedCodexInstalledVersion(npmPrefix: string): string | null {
+	const packageJsonPath = join(
+		npmPrefix,
+		"lib",
+		"node_modules",
+		"@openai",
+		"codex",
+		"package.json",
+	);
+	try {
+		const parsed = JSON.parse(readFileSync(packageJsonPath, "utf8")) as unknown;
+		if (!parsed || typeof parsed !== "object" || !("version" in parsed)) return null;
+		return typeof parsed.version === "string" && isValidSemver(parsed.version)
+			? parsed.version
+			: null;
+	} catch {
+		return null;
+	}
+}
+function installHostedCodexBootstrap(
+	packageSpec: string,
+	npmPrefix: string,
+	paths: RuntimePaths,
+): void {
+	if (!commandExists("npm")) {
+		throw new Error("Codex bootstrap requires npm on PATH");
+	}
+	mkdirSync(npmPrefix, { recursive: true });
+	const result = spawnRuntimeUserCommand(
+		"npm",
+		[
+			"install",
+			"-g",
+			"--prefix",
+			npmPrefix,
+			"--ignore-scripts",
+			"--fetch-retries",
+			"2",
+			"--fetch-retry-mintimeout",
+			"1000",
+			"--fetch-retry-maxtimeout",
+			"10000",
+			"--fetch-timeout",
+			"60000",
+			"--omit=dev",
+			"--no-audit",
+			"--no-fund",
+			"--no-update-notifier",
+			packageSpec,
+		],
+		paths.userHome,
+		paths.userHome,
+		{ timeoutMs: CODEX_BOOTSTRAP_TIMEOUT_MS },
+	);
+	if (result.status !== 0) {
+		throw new Error(
+			`Codex bootstrap failed: ${tail(result.stderr?.toString()) ?? tail(result.stdout?.toString()) ?? "npm failed"}`,
+		);
+	}
+}
+
+function quoteTomlString(value: string): string {
+	return JSON.stringify(value);
+}
