+++ packages/cli/src/runtime/manifest-providers.ts
@@ -1,141 +1,248 @@
-import { mkdirSync, readFileSync } from "node:fs";
-import { join } from "node:path";
-import {
-	isClawdiManagedProviderId,
-	MANAGED_AI_PROVIDER_RUNTIME_ENV,
-	nativeAiProvider,
-	nativeAiProviderRuntime,
-} from "@clawdi/shared";
-import { parse as parseYaml, stringify as stringifyYaml } from "yaml";
+import { parse as parseYaml } from "yaml";
 import { buildAgentTargetProjection } from "../lib/ai-provider-projection";
-import { writePrivateFileAtomic } from "../lib/private-file";
-import { isValidSemver } from "../lib/semver";
 import {
-	getHermesRawConfigValue,
-	getHermesResolvedConfigValue,
-	type HermesConfigTransaction,
-	reconcileHermesConfigValue,
-} from "./hermes-config";
-import { reconcileHermesNativeCredentials } from "./hermes-native-credentials";
-import { normalizeSecretRef } from "./hosted-egress-profiles";
+	applyHostedHermesAiProviderProjection,
+	buildOpenClawHostedProviderPatch,
+	type CatalogProviderConfigurationResult,
+	providerProjectionProgramImpact,
+} from "./catalog-provider-config";
+import type { HermesConfigTransaction } from "./hermes-config";
+import { applyHermesNativeProviders } from "./hermes-native-provider";
 import type { OpenClawHostedContext } from "./hosted-openclaw-context";
 import {
-	agentTargetProjectionInput,
-	type HostedAiProviderProjectionInput,
-	hostedAiProviderCatalog,
+	hostedProviderConfiguration,
+	hostedProviderEnvironment,
 	hostedProviderRequiresApiKey,
 } from "./hosted-provider-resolution";
 import type { RuntimeManifest } from "./manifest-contract";
-import {
-	type RuntimeInstallObservation,
-	runtimeFileCurrentRevision,
-	tail,
-} from "./manifest-install";
+import type { RuntimeInstallObservation } from "./manifest-install";
 import { removeOpenClawManagedProviderAuthProfiles } from "./manifest-oauth";
-import { canonicalJsonEqual, isPlainRecord, recordValue, stringValue } from "./manifest-shared";
-import { openClawPluginCapabilityConsentArgs } from "./openclaw-plugin-cli";
-import { openClawPluginListSchema } from "./openclaw-plugin-observation";
-import type { RuntimePaths } from "./paths";
-import { runtimeImpactRevision } from "./runtime-impact-revision";
+import { recordValue, stringValue } from "./manifest-shared";
 import {
-	commandExists,
-	executableExists,
-	runRuntimeUserCommand,
-	spawnRuntimeUserCommand,
-} from "./runtime-user-command";
+	applyOpenClawNativeProviders,
+	buildNativeOpenClawProviderPatch,
+	ensureNativeOpenClawProviderPlugins,
+} from "./openclaw-native-provider";
+import {
+	applyOpenClawGatewayHostedProjection,
+	applyOpenClawHostedProviderPatch,
+} from "./openclaw-provider-config";
+import { runtimeImpactRevision } from "./runtime-impact-revision";
 import { runtimeSecretValue } from "./secret-values";
 
-const CODEX_BOOTSTRAP_TIMEOUT_MS = 600_000;
-function ensureNativeOpenClawProviderPlugins(
-	input: HostedAiProviderProjectionInput | null,
-	commandPath: string,
+export { buildOpenClawHostedProviderPatch } from "./catalog-provider-config";
+export {
+	applyHostedCodexManagedProviderProjection,
+	CODEX_MANAGED_PROVIDER_CONFIG_FILE,
+	ensureHostedCodexCli,
+	hostedCodexHome,
+	hostedCodexManagedConfigToml,
+	hostedCodexManagedProvider,
+} from "./managed-codex-provider";
+export {
+	type OpenClawHostedProviderPatch,
+	openClawConfigPatchIsApplied,
+	openClawGatewayHostedPatch,
+} from "./openclaw-provider-config";
+
+interface HostedAiProviderProjectionResult extends CatalogProviderConfigurationResult {
+	nativeCredentialsChanged?: boolean;
+	nativeCredentialProviderIds?: string[];
+}
+
+export function applyHostedAiProviderProjection(
+	name: string,
+	observation: RuntimeInstallObservation,
+	manifest: RuntimeManifest,
+	secretValues: Record<string, string> | undefined,
 	home: string,
+	openClawContext: OpenClawHostedContext,
 	workspaceRoot: string,
-): void {
-	const required = new Map(
-		input?.catalog.providers.flatMap((provider) => {
-			const native =
-				provider.configuration_mode === "native"
-					? nativeAiProvider(provider.native_provider, provider.native_variant)
-					: undefined;
-			if (!native) return [];
-			const plugins = [[native.openclaw.plugin, native.openclaw.package] as const];
-			if (native.openclaw.companion_plugin && native.openclaw.companion_package)
-				plugins.push([native.openclaw.companion_plugin, native.openclaw.companion_package]);
-			return plugins;
-		}) ?? [],
-	);
-	if (required.size === 0) return;
-	const run = (args: string[]) => {
-		const result = spawnRuntimeUserCommand(commandPath, args, home, workspaceRoot, {
-			timeoutMs: 30_000,
-		});
+	previousProviderIds: readonly string[],
+	openClawOwnerBrowserBootstrapSupported: boolean,
+	hermesConfig: HermesConfigTransaction | null,
+	providerRevision: string,
+	previousNativeProviderIds: readonly string[] = [],
+): HostedAiProviderProjectionResult {
+	if (!observation.enabled || observation.status === "install_failed" || !observation.commandPath)
+		return { path: null, revision: null, providerIds: [] };
+	const { native, catalog } = hostedProviderConfiguration(manifest, name);
+	if (manifest.runtimes[name]?.providerMode === "configured" && !catalog && native.length === 0) {
+		if (name === "openclaw")
+			applyOpenClawGatewayHostedProjection(
+				observation.commandPath,
+				manifest,
+				secretValues,
+				openClawContext,
+				workspaceRoot,
+				openClawOwnerBrowserBootstrapSupported,
+			);
 		return {
-			status: result.status,
-			stdout: String(result.stdout ?? ""),
-			stderr: String(result.stderr ?? ""),
+			path: null,
+			revision: null,
+			providerIds: [...previousProviderIds],
+			nativeCredentialProviderIds: [...previousNativeProviderIds],
 		};
-	};
-	const observe = () => {
-		const result = run(["plugins", "list", "--json"]);
-		if (result.status !== 0)
-			throw new Error("Native OpenClaw provider plugins could not be inspected");
-		return openClawPluginListSchema.parse(JSON.parse(String(result.stdout)));
-	};
-	let installed = observe();
-	for (const [plugin, packageSpec] of required) {
-		if (!installed.plugins.some((entry) => entry.id === plugin)) {
-			// The target is absent. --force acknowledges this fixed official npm
-			// source; native security policy still governs installation.
-			runRuntimeUserCommand(
-				commandPath,
-				[
-					"plugins",
-					"install",
-					packageSpec,
-					"--force",
-					"--pin",
-					...openClawPluginCapabilityConsentArgs("install", run),
-				],
-				"",
-				home,
+	}
+	if (name === "hermes") {
+		if (!hermesConfig) throw new Error("Hermes config command is unavailable");
+		const auth = applyHermesNativeProviders({
+			connections: native,
+			previousProviderIds: previousNativeProviderIds,
+			secretValues: secretValues ?? {},
+			config: hermesConfig,
+			home,
+			workspaceRoot,
+		});
+		const projection = applyHostedHermesAiProviderProjection(
+			catalog,
+			previousProviderIds,
+			home,
+			hermesConfig,
+			true,
+			native.length > 0,
+		);
+		return {
+			...projection,
+			nativeCredentialsChanged: auth.changed,
+			nativeCredentialProviderIds: auth.providerIds,
+		};
+	}
+	if (name !== "openclaw") return { path: null, revision: null, providerIds: [] };
+	const patch = buildOpenClawHostedProviderPatch(
+		catalog,
+		previousProviderIds,
+		native.length > 0 ? "merge" : "replace",
+	);
+	const { placeholderEnv, configEnv, secretEnv } = hostedProviderEnvironment(manifest, name);
+	const environment = { ...placeholderEnv, ...configEnv };
+	for (const [key, ref] of Object.entries(secretEnv)) {
+		const value = runtimeSecretValue(secretValues ?? {}, ref);
+		if (!value) throw new Error("OpenClaw provider credential is unavailable");
+		environment[key] = value;
+	}
+	let nativeChanged = ensureNativeOpenClawProviderPlugins(
+		native,
+		observation.commandPath,
+		openClawContext.home,
+		workspaceRoot,
+		environment,
+	);
+	applyOpenClawGatewayHostedProjection(
+		observation.commandPath,
+		manifest,
+		secretValues,
+		openClawContext,
+		workspaceRoot,
+		openClawOwnerBrowserBootstrapSupported,
+		environment,
+	);
+	// The catalog SDK owns exact replacement and large shrink. It runs first so
+	// migration from a whole-owned catalog row to native auth cannot retain models.
+	if (patch.apply)
+		applyOpenClawHostedProviderPatch(
+			patch,
+			observation.commandPath,
+			openClawContext,
+			workspaceRoot,
+			providerRevision,
+		);
+	const nativePatch = buildNativeOpenClawProviderPatch(
+		native,
+		previousNativeProviderIds.filter(
+			(id) => !patch.providerIds.includes(id) && !previousProviderIds.includes(id),
+		),
+	);
+	if (native.length > 0 || previousNativeProviderIds.length > 0) {
+		nativeChanged =
+			applyOpenClawNativeProviders({
+				patch: nativePatch,
+				command: observation.commandPath,
+				context: openClawContext,
 				workspaceRoot,
-				{ timeoutMs: 180_000 },
-			);
-			installed = observe();
-		}
-		const matches = installed.plugins.filter((entry) => entry.id === plugin);
-		if (matches.length !== 1 || matches[0]?.status === "error") {
-			throw new Error(`Required native OpenClaw provider plugin ${plugin} is unavailable`);
-		}
-		if (matches[0]?.enabled !== true) {
-			runRuntimeUserCommand(
-				commandPath,
-				["plugins", "enable", plugin, ...openClawPluginCapabilityConsentArgs("enable", run)],
-				"",
-				home,
-				workspaceRoot,
-				{ timeoutMs: 30_000 },
-			);
-			installed = observe();
-		}
-		if (
-			!installed.plugins.some(
-				(entry) => entry.id === plugin && entry.enabled && entry.status === "loaded",
-			)
-		)
-			throw new Error(`Required native OpenClaw provider plugin ${plugin} did not activate`);
+				environment,
+			}) || nativeChanged;
+	}
+	if (openClawContext.managedApiKeyProjection) {
+		if (openClawContext.agentDirs.managed.length === 0)
+			throw new Error("OpenClaw managed provider-auth stores were not transactionally discovered");
+		removeOpenClawManagedProviderAuthProfiles(openClawContext, workspaceRoot, providerRevision);
+	}
+	return {
+		path: observation.commandPath,
+		revision: null,
+		providerIds: patch.providerIds,
+		nativeCredentialProviderIds: nativePatch.providerIds,
+		nativeCredentialsChanged: nativeChanged,
+	};
+}
+
+export function previewHostedAiProviderProjectionRevision(
+	name: string,
+	observation: RuntimeInstallObservation,
+	manifest: RuntimeManifest,
+	home: string,
+	previousProviderIds: readonly string[],
+	previousNativeProviderIds: readonly string[] = [],
+): string | null {
+	if (
+		(name !== "openclaw" && name !== "hermes") ||
+		!observation.enabled ||
+		observation.status === "install_failed" ||
+		!observation.commandPath
+	)
+		return null;
+	const { native, catalog } = hostedProviderConfiguration(manifest, name);
+	if (manifest.runtimes[name]?.providerMode === "configured" && !catalog && native.length === 0)
+		return null;
+	if (name === "hermes")
+		return applyHostedHermesAiProviderProjection(
+			catalog,
+			previousProviderIds,
+			home,
+			null,
+			false,
+			native.length > 0,
+		).revision;
+	const patch = buildOpenClawHostedProviderPatch(
+		catalog,
+		previousProviderIds,
+		native.length > 0 ? "merge" : "replace",
+	);
+	return runtimeImpactRevision({
+		catalog: providerProjectionProgramImpact("openclaw", JSON.parse(patch.content), catalog),
+		native: buildNativeOpenClawProviderPatch(
+			native,
+			previousNativeProviderIds.filter(
+				(id) => !patch.providerIds.includes(id) && !previousProviderIds.includes(id),
+			),
+		).config,
+	});
+}
+
+/** Validate the selected path without materializing any credentials. */
+export function validateHostedProviderConfiguration(
+	manifest: RuntimeManifest,
+	name: string,
+	previousProviderIds: readonly string[],
+): void {
+	const { native, catalog } = hostedProviderConfiguration(manifest, name);
+	if (name === "openclaw") {
+		buildOpenClawHostedProviderPatch(
+			catalog,
+			previousProviderIds,
+			native.length > 0 ? "merge" : "replace",
+		);
+		if (native.length > 0) buildNativeOpenClawProviderPatch(native, []);
+	} else if (name === "hermes" && catalog) {
+		const projection = buildAgentTargetProjection("hermes", catalog.catalog, catalog.primaryModel, {
+			freezeManagedModelCatalog: true,
+		});
+		const file = projection.files.find((entry) => entry.path.endsWith(".hermes.yaml"));
+		if (!file || !recordValue(parseYaml(file.content)))
+			throw new Error("Hermes provider projection must be a YAML object");
 	}
 }
-const OPENCLAW_SCHEMA_PROBE_TIMEOUT_MS = 15_000;
-const OPENCLAW_SCHEMA_PROBE_MAX_BYTES = 16 * 1024 * 1024;
-const openClawProviderPatchRevisions = new Map<string, string>();
-const openClawMemorySearchLayouts = new Map<
-	string,
-	{ revision: string; layout: OpenClawMemorySearchLayout }
->();
-
-type OpenClawMemorySearchLayout = "agents-defaults" | "top-level";
 
 export function providerHealthReasons(
 	provider: Record<string, unknown>,
@@ -198,1093 +305,4 @@ function providerHasModels(provider: Record<string, unknown>): boolean {
 }
 function isOpenAiCompatibleMode(apiMode: string | null): boolean {
 	return apiMode === "openai_chat" || apiMode === "openai_responses";
-}
-interface HostedAiProviderProjectionResult {
-	nativeCredentialsChanged?: boolean;
-	nativeCredentialProviderIds?: string[];
-	path: string | null;
-	revision: string | null;
-	providerIds: string[];
-}
-const CODEX_MANAGED_PROVIDER_ID = "clawdi";
-export const CODEX_MANAGED_PROVIDER_CONFIG_FILE = "config.toml";
-const CODEX_BOOTSTRAP_PACKAGE_VERSION = "0.146.0";
-const CODEX_BOOTSTRAP_PACKAGE_SPEC = `@openai/codex@${CODEX_BOOTSTRAP_PACKAGE_VERSION}`;
-interface HostedCodexManagedProvider {
-	baseUrl: string;
-}
-export function applyHostedAiProviderProjection(
-	name: string,
-	observation: RuntimeInstallObservation,
-	manifest: RuntimeManifest,
-	secretValues: Record<string, string> | undefined,
-	home: string,
-	openClawContext: OpenClawHostedContext,
-	workspaceRoot: string,
-	previousProviderIds: readonly string[],
-	openClawOwnerBrowserBootstrapSupported: boolean,
-	hermesConfig: HermesConfigTransaction | null,
-	providerRevision: string,
-	previousNativeProviderIds: readonly string[] = [],
-): HostedAiProviderProjectionResult {
-	if (!observation.enabled || observation.status === "install_failed" || !observation.commandPath) {
-		return { path: null, revision: null, providerIds: [] };
-	}
-	const projectionInput = agentTargetProjectionInput(hostedAiProviderCatalog(manifest, name));
-	assertHostedProviderProjectionMode(name, manifest, projectionInput);
-	if (manifest.runtimes[name]?.providerMode === "configured" && !projectionInput) {
-		if (name === "openclaw") {
-			applyOpenClawGatewayHostedProjection(
-				observation.commandPath,
-				manifest,
-				secretValues,
-				openClawContext,
-				workspaceRoot,
-				openClawOwnerBrowserBootstrapSupported,
-			);
-		}
-		return {
-			path: null,
-			revision: null,
-			providerIds: [...previousProviderIds],
-			nativeCredentialProviderIds: [...previousNativeProviderIds],
-		};
-	}
-	if (name === "hermes") {
-		if (!hermesConfig) throw new Error("Hermes config command is unavailable");
-		const providers =
-			projectionInput?.catalog.providers.flatMap((provider) => {
-				if (provider.configuration_mode !== "native" || provider.auth.type === "agent_profile")
-					return [];
-				const native = nativeAiProvider(provider.native_provider, provider.native_variant);
-				if (!native) throw new Error("Invalid native Hermes provider identity");
-				const ref = manifest.projection?.providers?.[provider.id]?.apiKeySecretRef;
-				const apiKey = ref ? runtimeSecretValue(secretValues ?? {}, ref) : null;
-				if (!apiKey) throw new Error("Native Hermes provider credential is unavailable");
-				return [
-					{
-						providerId: native.hermes.provider,
-						apiKey,
-						baseUrl: nativeAiProviderRuntime(native, "hermes").base_url,
-					},
-				];
-			}) ?? [];
-		const currentStrategies = getHermesRawConfigValue(hermesConfig, "credential_pool_strategies");
-		if (currentStrategies.exists && !isPlainRecord(currentStrategies.value))
-			throw new Error("Hermes credential strategies must be an object");
-		const strategies = isPlainRecord(currentStrategies.value) ? currentStrategies.value : {};
-		const selected = getHermesRawConfigValue(hermesConfig, "model.provider");
-		const auth = reconcileHermesNativeCredentials({
-			home,
-			workspaceRoot,
-			providers,
-			previousProviderIds: previousNativeProviderIds,
-			strategies,
-			selectedProvider: stringValue(selected.value)?.trim().toLowerCase(),
-		});
-		const ownsSelectedConnection = projectionInput?.catalog.providers.some(
-			(provider) =>
-				provider.configuration_mode === "native" &&
-				nativeAiProvider(provider.native_provider, provider.native_variant)?.hermes.provider ===
-					auth.selectedProvider,
-		);
-		let connectionOverridesChanged = false;
-		if (ownsSelectedConnection) {
-			for (const field of HERMES_DIRECT_MODEL_FIELDS) {
-				const path = `model.${field}`;
-				if (getHermesRawConfigValue(hermesConfig, path).exists) {
-					reconcileHermesConfigValue(hermesConfig, path, undefined);
-					connectionOverridesChanged = true;
-				}
-			}
-		}
-		if (Object.keys(auth.strategyUpdates).length > 0) {
-			const next = { ...strategies };
-			for (const [providerId, strategy] of Object.entries(auth.strategyUpdates)) {
-				if (strategy.exists) next[providerId] = strategy.value;
-				else delete next[providerId];
-			}
-			reconcileHermesConfigValue(
-				hermesConfig,
-				"credential_pool_strategies",
-				Object.keys(next).length > 0 ? next : undefined,
-			);
-		}
-		const projection = applyHostedHermesAiProviderProjection(
-			projectionInput,
-			previousProviderIds,
-			home,
-			hermesConfig,
-		);
-		return {
-			...projection,
-			nativeCredentialsChanged: auth.changed || connectionOverridesChanged,
-			nativeCredentialProviderIds: providers.map((provider) => provider.providerId),
-		};
-	}
-	if (name === "openclaw") {
-		ensureNativeOpenClawProviderPlugins(
-			projectionInput,
-			observation.commandPath,
-			openClawContext.home,
-			workspaceRoot,
-		);
-		applyOpenClawGatewayHostedProjection(
-			observation.commandPath,
-			manifest,
-			secretValues,
-			openClawContext,
-			workspaceRoot,
-			openClawOwnerBrowserBootstrapSupported,
-		);
-		const providerPatch = buildOpenClawHostedProviderPatch(
-			projectionInput,
-			previousProviderIds,
-			previousNativeProviderIds,
-		);
-		if (providerPatch.apply) {
-			applyOpenClawHostedProviderPatch(
-				providerPatch,
-				observation.commandPath,
-				openClawContext,
-				workspaceRoot,
-				providerRevision,
-				[
-					...previousNativeProviderIds,
-					...(projectionInput?.catalog.providers
-						.filter((provider) => provider.configuration_mode === "native")
-						.flatMap((provider) => {
-							const native = nativeAiProvider(provider.native_provider, provider.native_variant);
-							return native && !previousProviderIds.includes(native.openclaw.provider)
-								? [native.openclaw.provider]
-								: [];
-						}) ?? []),
-				],
-			);
-		}
-		if (openClawContext.managedApiKeyProjection) {
-			if (openClawContext.agentDirs.managed.length === 0) {
-				throw new Error(
-					"OpenClaw managed provider-auth stores were not transactionally discovered",
-				);
-			}
-			removeOpenClawManagedProviderAuthProfiles(openClawContext, workspaceRoot, providerRevision);
-		}
-		return {
-			path: observation.commandPath,
-			revision: null,
-			providerIds: providerPatch.providerIds,
-			nativeCredentialProviderIds: providerPatch.nativeCredentialProviderIds,
-		};
-	}
-	return { path: null, revision: null, providerIds: [] };
-}
-export function previewHostedAiProviderProjectionRevision(
-	name: string,
-	observation: RuntimeInstallObservation,
-	manifest: RuntimeManifest,
-	home: string,
-	previousProviderIds: readonly string[],
-	previousNativeProviderIds: readonly string[] = [],
-): string | null {
-	if (
-		(name !== "openclaw" && name !== "hermes") ||
-		!observation.enabled ||
-		observation.status === "install_failed" ||
-		!observation.commandPath
-	) {
-		return null;
-	}
-	const projectionInput = agentTargetProjectionInput(hostedAiProviderCatalog(manifest, name));
-	assertHostedProviderProjectionMode(name, manifest, projectionInput);
-	if (manifest.runtimes[name]?.providerMode === "configured" && !projectionInput) {
-		return null;
-	}
-	if (name === "openclaw") {
-		const providerPatch = buildOpenClawHostedProviderPatch(
-			projectionInput,
-			previousProviderIds,
-			previousNativeProviderIds,
-		);
-		if (!projectionInput) {
-			return runtimeImpactRevision({
-				openClawProviderProjection: "delete",
-				patch: JSON.parse(providerPatch.content) as unknown,
-			});
-		}
-		return runtimeImpactRevision({
-			openClawProviderProjection: "json-patch",
-			patch: providerProjectionProgramImpact(
-				"openclaw",
-				JSON.parse(providerPatch.content) as unknown,
-				projectionInput,
-			),
-		});
-	}
-	return applyHostedHermesAiProviderProjection(
-		projectionInput,
-		previousProviderIds,
-		home,
-		null,
-		false,
-	).revision;
-}
-function providerProjectionProgramImpact(
-	runtime: "openclaw" | "hermes",
-	patch: unknown,
-	projectionInput: HostedAiProviderProjectionInput,
-): unknown {
-	const root = recordValue(patch);
-	const managedProviderIds = new Set(
-		projectionInput.catalog.providers
-			.filter((provider) => provider.managed_by === "clawdi")
-			.map((provider) => provider.id),
-	);
-	if (!root || managedProviderIds.size === 0) return patch;
-
-	const providerContainer = runtime === "openclaw" ? recordValue(root.models) : root;
-	if (!providerContainer) return patch;
-	const providers = recordValue(providerContainer.providers);
-	if (!providers) return patch;
-	const programProviders = Object.fromEntries(
-		Object.entries(providers).map(([providerId, provider]) => {
-			const providerConfig = recordValue(provider);
-			if (!managedProviderIds.has(providerId) || !providerConfig) return [providerId, provider];
-			const { models: _models, ...programConfig } = providerConfig;
-			return [providerId, programConfig];
-		}),
-	);
-	if (runtime === "openclaw") {
-		return { ...root, models: { ...providerContainer, providers: programProviders } };
-	}
-	return { ...root, providers: programProviders };
-}
-export function applyHostedCodexManagedProviderProjection(
-	manifest: RuntimeManifest,
-	home: string,
-	codexCli: Record<string, string> | null,
-): HostedAiProviderProjectionResult {
-	const provider = hostedCodexManagedProvider(manifest);
-	if (!provider) return { path: null, revision: null, providerIds: [] };
-
-	const codexHome = hostedCodexHome(home);
-	const configPath = join(codexHome, CODEX_MANAGED_PROVIDER_CONFIG_FILE);
-	const configContent = hostedCodexManagedConfigToml(provider);
-	writePrivateFileAtomic(configPath, configContent, { mode: 0o600, dirMode: 0o700 });
-
-	return {
-		path: configPath,
-		providerIds: [CODEX_MANAGED_PROVIDER_ID],
-		revision: runtimeImpactRevision({
-			codexManagedProviderProjection: CODEX_MANAGED_PROVIDER_CONFIG_FILE,
-			configContent,
-			codexCli,
-		}),
-	};
-}
-export function assertHostedProviderProjectionMode(
-	runtimeName: string,
-	manifest: RuntimeManifest,
-	projectionInput: HostedAiProviderProjectionInput | null,
-): void {
-	const providerMode = manifest.runtimes[runtimeName]?.providerMode;
-	if (providerMode === "unmanaged" && projectionInput) {
-		throw new Error(`runtime ${runtimeName} unmanaged provider mode has a provider projection`);
-	}
-}
-export function hostedCodexManagedProvider(
-	manifest: RuntimeManifest,
-): HostedCodexManagedProvider | null {
-	const terminalTooling = recordValue(manifest.projection?.terminalTooling);
-	const codex = recordValue(terminalTooling?.codex);
-	const provider = recordValue(codex?.provider);
-	const primaryModel = recordValue(codex?.primary_model);
-	const providerId = stringValue(codex?.provider_id);
-	const baseUrl = stringValue(provider?.baseUrl);
-	const apiMode = stringValue(provider?.apiMode);
-	const primaryModelName = stringValue(primaryModel?.model);
-	if (
-		codex?.enabled !== true ||
-		!provider ||
-		provider.managed_by !== "clawdi" ||
-		apiMode !== "openai_responses" ||
-		stringValue(provider.runtimeEnvName) !== MANAGED_AI_PROVIDER_RUNTIME_ENV ||
-		normalizeSecretRef(stringValue(provider.apiKeySecretRef)) !== "secret://tool.codex.apiKey" ||
-		!providerId ||
-		stringValue(primaryModel?.provider_id) !== providerId ||
-		!baseUrl ||
-		!primaryModelName
-	) {
-		return null;
-	}
-	return { baseUrl };
-}
-export function hostedCodexHome(home: string): string {
-	return join(home, ".codex");
-}
-export function hostedCodexManagedConfigToml(provider: HostedCodexManagedProvider): string {
-	const lines = ["# Generated by Clawdi hosted runtime. Do not put API keys in this file."];
-	lines.push(
-		`model_provider = ${quoteTomlString(CODEX_MANAGED_PROVIDER_ID)}`,
-		"",
-		`[model_providers.${CODEX_MANAGED_PROVIDER_ID}]`,
-		`name = ${quoteTomlString("clawdi")}`,
-		`base_url = ${quoteTomlString(provider.baseUrl)}`,
-		`env_key = ${quoteTomlString(MANAGED_AI_PROVIDER_RUNTIME_ENV)}`,
-		'wire_api = "responses"',
-		"",
-	);
-	return lines.join("\n");
-}
-export function ensureHostedCodexCli(paths: RuntimePaths): Record<string, string> | null {
-	if (process.env.CLAWDI_CODEX_INSTALL_DISABLED === "1") return null;
-	const npmPrefix = paths.userNpmPrefix;
-	const realBin = join(npmPrefix, "bin", "codex");
-	let installedVersion = hostedCodexInstalledVersion(npmPrefix);
-	const bootstrapRequired = installedVersion === null || !executableExists(realBin);
-	if (bootstrapRequired) {
-		installHostedCodexBootstrap(CODEX_BOOTSTRAP_PACKAGE_SPEC, npmPrefix, paths);
-		installedVersion = hostedCodexInstalledVersion(npmPrefix);
-		if (installedVersion !== CODEX_BOOTSTRAP_PACKAGE_VERSION) {
-			throw new Error(
-				`Codex bootstrap installed version ${installedVersion ?? "unknown"}; expected ${CODEX_BOOTSTRAP_PACKAGE_VERSION}`,
-			);
-		}
-		if (!executableExists(realBin)) {
-			throw new Error(`Codex bootstrap did not create ${realBin}`);
-		}
-	}
-	if (installedVersion === null) throw new Error("Codex package metadata is unavailable");
-	return {
-		commandPath: realBin,
-		npmPrefix,
-		bootstrapPackageSpec: CODEX_BOOTSTRAP_PACKAGE_SPEC,
-		installedVersion,
-		realBin,
-	};
-}
-function hostedCodexInstalledVersion(npmPrefix: string): string | null {
-	const packageJsonPath = join(
-		npmPrefix,
-		"lib",
-		"node_modules",
-		"@openai",
-		"codex",
-		"package.json",
-	);
-	try {
-		const parsed = JSON.parse(readFileSync(packageJsonPath, "utf8")) as unknown;
-		if (!parsed || typeof parsed !== "object" || !("version" in parsed)) return null;
-		return typeof parsed.version === "string" && isValidSemver(parsed.version)
-			? parsed.version
-			: null;
-	} catch {
-		return null;
-	}
-}
-function installHostedCodexBootstrap(
-	packageSpec: string,
-	npmPrefix: string,
-	paths: RuntimePaths,
-): void {
-	if (!commandExists("npm")) {
-		throw new Error("Codex bootstrap requires npm on PATH");
-	}
-	mkdirSync(npmPrefix, { recursive: true });
-	const result = spawnRuntimeUserCommand(
-		"npm",
-		[
-			"install",
-			"-g",
-			"--prefix",
-			npmPrefix,
-			"--ignore-scripts",
-			"--fetch-retries",
-			"2",
-			"--fetch-retry-mintimeout",
-			"1000",
-			"--fetch-retry-maxtimeout",
-			"10000",
-			"--fetch-timeout",
-			"60000",
-			"--omit=dev",
-			"--no-audit",
-			"--no-fund",
-			"--no-update-notifier",
-			packageSpec,
-		],
-		paths.userHome,
-		paths.userHome,
-		{ timeoutMs: CODEX_BOOTSTRAP_TIMEOUT_MS },
-	);
-	if (result.status !== 0) {
-		throw new Error(
-			`Codex bootstrap failed: ${tail(result.stderr?.toString()) ?? tail(result.stdout?.toString()) ?? "npm failed"}`,
-		);
-	}
-}
-function applyHostedHermesAiProviderProjection(
-	projectionInput: HostedAiProviderProjectionInput | null,
-	previousProviderIds: readonly string[],
-	home: string,
-	config: HermesConfigTransaction | null,
-	apply = true,
-): HostedAiProviderProjectionResult {
-	const configPath = join(home, ".hermes", "config.yaml");
-	if (!projectionInput) {
-		const deletedProviderIds = staleProviderIds(new Set(previousProviderIds), new Set());
-		if (apply && deletedProviderIds.length > 0) {
-			if (!config) throw new Error("Hermes config command is unavailable");
-			applyHermesProviderConfig(config, {}, deletedProviderIds);
-		}
-		return {
-			path: null,
-			providerIds: [],
-			revision: runtimeImpactRevision({
-				hermesProviderProjection: "none",
-				deletedProviderIds,
-			}),
-		};
-	}
-
-	const projection = buildAgentTargetProjection(
-		"hermes",
-		projectionInput.catalog,
-		projectionInput.primaryModel,
-		{ freezeManagedModelCatalog: true },
-	);
-	const file = projection.files.find((entry) => entry.path.endsWith(".hermes.yaml"));
-	if (!file) throw new Error("Hermes projection did not include a config merge YAML file.");
-	const activeProviderIds = [...providerIdsFromPatch("hermes", file.content)].sort();
-	const deletedProviderIds = staleProviderIds(
-		new Set(previousProviderIds),
-		new Set(activeProviderIds),
-	);
-	const patchContent = mergeProviderDeletes("hermes", file.content, deletedProviderIds);
-	if (apply) {
-		if (!config) throw new Error("Hermes config command is unavailable");
-		const patch = parseYaml(file.content) as unknown;
-		const root = recordValue(patch);
-		if (!root) throw new Error("Hermes projection patch must be a YAML object.");
-		applyHermesProviderConfig(
-			config,
-			root,
-			deletedProviderIds,
-			projectionInput.catalog.providers.some(
-				(provider) => provider.configuration_mode === "native",
-			),
-		);
-	}
-	return {
-		path: configPath,
-		providerIds: activeProviderIds,
-		revision: runtimeImpactRevision({
-			hermesProviderProjection: "yaml-merge",
-			patch: providerProjectionProgramImpact("hermes", parseYaml(patchContent), projectionInput),
-		}),
-	};
-}
-function quoteTomlString(value: string): string {
-	return JSON.stringify(value);
-}
-export interface OpenClawHostedProviderPatch {
-	nativeCredentialProviderIds?: string[];
-	apply: boolean;
-	content: string;
-	providerIds: string[];
-}
-const OPENCLAW_CONFIG_MUTATION_HELPER = `
-import { readFileSync } from "node:fs";
-import { pathToFileURL } from "node:url";
-import { isDeepStrictEqual } from "node:util";
-
-const sdk = await import(pathToFileURL(process.argv[1]).href);
-if (
-  typeof sdk.readConfigFileSnapshotForWrite !== "function" ||
-  typeof sdk.mutateConfigFile !== "function"
-) {
-  throw new Error("required public config-mutation export is missing");
-}
-const patch = JSON.parse(readFileSync(0, "utf8"));
-const nativeProviderIds = new Set(JSON.parse(process.argv[2]));
-const isRecord = (value) => value !== null && typeof value === "object" && !Array.isArray(value);
-if (!isRecord(patch)) throw new Error("OpenClaw provider patch must be an object");
-const blockedKeys = new Set(["__proto__", "constructor", "prototype"]);
-const explicitSetPaths = [];
-const unsetPaths = [];
-const applyMergePatch = (target, source, path = []) => {
-  for (const [key, value] of Object.entries(source)) {
-    if (blockedKeys.has(key)) throw new Error("OpenClaw provider patch contains a blocked key");
-    const nextPath = [...path, key];
-    if (value === null) {
-      delete target[key];
-      unsetPaths.push(nextPath);
-    } else if (path.length === 2 && path[0] === "models" && path[1] === "providers" && !nativeProviderIds.has(key)) {
-      target[key] = structuredClone(value);
-      explicitSetPaths.push(nextPath);
-    } else if (isRecord(value)) {
-      if (!isRecord(target[key])) target[key] = {};
-      if (Object.keys(value).length === 0) explicitSetPaths.push(nextPath);
-      applyMergePatch(target[key], value, nextPath);
-      if (path.length === 2 && path[0] === "models" && path[1] === "providers" && nativeProviderIds.has(key) && Object.keys(target[key]).length === 0) {
-        delete target[key];
-        unsetPaths.push(nextPath);
-      }
-    } else {
-      target[key] = structuredClone(value);
-      explicitSetPaths.push(nextPath);
-    }
-  }
-};
-const configRead = await sdk.readConfigFileSnapshotForWrite({ skipPluginValidation: true });
-const snapshot = configRead?.snapshot;
-const sourceConfig = snapshot?.sourceConfig;
-const sourceAgents = isRecord(sourceConfig) ? sourceConfig.agents : undefined;
-const sourceDefaults = isRecord(sourceAgents) ? sourceAgents.defaults : undefined;
-const sourceMemory = isRecord(sourceConfig) ? sourceConfig.memory : undefined;
-const patchAgents = patch.agents;
-const patchDefaults = isRecord(patchAgents) ? patchAgents.defaults : undefined;
-const patchMemory = patch.memory;
-const repairsUnsupportedMemorySearch =
-  (isRecord(sourceDefaults) &&
-    Object.hasOwn(sourceDefaults, "memorySearch") &&
-    isRecord(patchDefaults) &&
-    patchDefaults.memorySearch === null) ||
-  (isRecord(sourceMemory) &&
-    Object.hasOwn(sourceMemory, "search") &&
-    isRecord(patchMemory) &&
-    patchMemory.search === null &&
-    isRecord(patchDefaults?.memorySearch));
-if (
-  !snapshot ||
-  !isRecord(sourceConfig) ||
-  (snapshot.valid !== true && !repairsUnsupportedMemorySearch)
-) {
-  throw new Error("OpenClaw config snapshot is unavailable for provider projection");
-}
-const projected = structuredClone(sourceConfig);
-applyMergePatch(projected, patch);
-if (isDeepStrictEqual(projected, sourceConfig)) process.exit(0);
-explicitSetPaths.length = 0;
-unsetPaths.length = 0;
-await sdk.mutateConfigFile({
-  base: "source",
-  afterWrite: { mode: "none", reason: "Clawdi runtime convergence owns service reconciliation" },
-  writeOptions: { allowConfigSizeDrop: true, explicitSetPaths, unsetPaths },
-  mutate: (draft) => applyMergePatch(draft, patch),
-});
-`;
-function applyOpenClawHostedProviderPatch(
-	patch: OpenClawHostedProviderPatch,
-	commandPath: string,
-	context: OpenClawHostedContext,
-	workspaceRoot: string,
-	providerRevision: string,
-	nativeProviderIds: string[] = [],
-): void {
-	const sdkPath = context.requireSdkExport("configMutation");
-	const content = adaptOpenClawMemorySearchPatch(
-		patch.content,
-		commandPath,
-		context.home,
-		workspaceRoot,
-		sdkPath,
-	);
-	const patchRevision = runtimeImpactRevision({
-		providerRevision,
-		content,
-		sdk: runtimeFileCurrentRevision(sdkPath),
-	});
-	if (openClawProviderPatchRevisions.get(context.configPath) === patchRevision) {
-		const expected = recordValue(JSON.parse(content) as unknown);
-		if (!expected) throw new Error("OpenClaw provider projection patch must be an object");
-		if (openClawConfigPatchIsApplied(context, expected)) return;
-	}
-	runRuntimeUserCommand(
-		"node",
-		[
-			"--input-type=module",
-			"--eval",
-			OPENCLAW_CONFIG_MUTATION_HELPER,
-			sdkPath,
-			JSON.stringify(nativeProviderIds),
-		],
-		content,
-		context.home,
-		workspaceRoot,
-	);
-	openClawProviderPatchRevisions.set(context.configPath, patchRevision);
-}
-
-function adaptOpenClawMemorySearchPatch(
-	content: string,
-	commandPath: string,
-	home: string,
-	workspaceRoot: string,
-	sdkPath: string,
-): string {
-	const root = recordValue(JSON.parse(content) as unknown);
-	if (!root) throw new Error("OpenClaw provider projection patch must be an object");
-	const memory = recordValue(root.memory);
-	if (!memory || !Object.hasOwn(memory, "search")) return content;
-	const search = memory.search;
-	if (openClawMemorySearchLayout(commandPath, home, workspaceRoot, sdkPath) === "top-level") {
-		return content;
-	}
-
-	const patch = { ...root };
-	const agents = { ...(recordValue(patch.agents) ?? {}) };
-	const defaults = { ...(recordValue(agents.defaults) ?? {}) };
-	defaults.memorySearch = search;
-	agents.defaults = defaults;
-	patch.agents = agents;
-	patch.memory = { ...memory, search: null };
-	return `${JSON.stringify(patch, null, 2)}\n`;
-}
-
-function openClawMemorySearchLayout(
-	commandPath: string,
-	home: string,
-	workspaceRoot: string,
-	sdkPath: string,
-): OpenClawMemorySearchLayout {
-	const revision = [
-		runtimeFileCurrentRevision(commandPath),
-		runtimeFileCurrentRevision(sdkPath),
-	].join("\0");
-	const cached = openClawMemorySearchLayouts.get(commandPath);
-	if (cached?.revision === revision) return cached.layout;
-
-	const result = spawnRuntimeUserCommand(commandPath, ["config", "schema"], home, workspaceRoot, {
-		timeoutMs: OPENCLAW_SCHEMA_PROBE_TIMEOUT_MS,
-		maxBufferBytes: OPENCLAW_SCHEMA_PROBE_MAX_BYTES,
-	});
-	if (result.status !== 0) {
-		throw new Error("installed OpenClaw config schema is unavailable");
-	}
-	let schema: unknown;
-	try {
-		schema = JSON.parse(String(result.stdout));
-	} catch {
-		throw new Error("installed OpenClaw config schema is malformed");
-	}
-	const topLevel = jsonSchemaHasPropertyPath(schema, ["memory", "search"]);
-	const agentsDefaults = jsonSchemaHasPropertyPath(schema, ["agents", "defaults", "memorySearch"]);
-	if (topLevel === agentsDefaults) {
-		throw new Error("installed OpenClaw memory search schema is unsupported");
-	}
-	const layout: OpenClawMemorySearchLayout = topLevel ? "top-level" : "agents-defaults";
-	openClawMemorySearchLayouts.set(commandPath, { revision, layout });
-	return layout;
-}
-
-function jsonSchemaHasPropertyPath(schema: unknown, path: readonly string[]): boolean {
-	let current = recordValue(schema);
-	for (const segment of path) {
-		const properties = current ? recordValue(current.properties) : null;
-		current = properties ? recordValue(properties[segment]) : null;
-		if (!current) return false;
-	}
-	return true;
-}
-export function buildOpenClawHostedProviderPatch(
-	projectionInput: HostedAiProviderProjectionInput | null,
-	previousProviderIds: readonly string[],
-	previousNativeProviderIds: readonly string[] = [],
-): OpenClawHostedProviderPatch {
-	if (!projectionInput) {
-		const deletedProviderIds = staleProviderIds(new Set(previousProviderIds), new Set());
-		return {
-			apply: deletedProviderIds.length > 0 || previousNativeProviderIds.length > 0,
-			content: removeNativeOpenClawCredentials(
-				mergeProviderDeletes("openclaw", "{}\n", deletedProviderIds),
-				previousNativeProviderIds,
-			),
-			providerIds: [],
-			nativeCredentialProviderIds: [],
-		};
-	}
-	const projection = buildAgentTargetProjection(
-		"openclaw",
-		projectionInput.catalog,
-		projectionInput.primaryModel,
-	);
-	const file = projection.files.find((entry) => entry.path.endsWith(".openclaw.json"));
-	if (!file) throw new Error("OpenClaw projection did not include a config patch JSON file.");
-	const projectedIds = providerIdsFromPatch("openclaw", file.content);
-	const nativeIds = new Set(
-		projectionInput.catalog.providers.flatMap((provider) => {
-			const native =
-				provider.configuration_mode === "native"
-					? nativeAiProvider(provider.native_provider, provider.native_variant)
-					: undefined;
-			return native && native.id !== "openai-codex" ? [native.openclaw.provider] : [];
-		}),
-	);
-	const providerIds = [...projectedIds].filter((id) => !nativeIds.has(id)).sort();
-	const deletedProviderIds = staleProviderIds(new Set(previousProviderIds), projectedIds);
-	let providerPatchContent = file.content;
-	if (
-		projectionInput.catalog.providers.some((provider) => provider.configuration_mode === "native")
-	) {
-		providerPatchContent = withOpenClawProviderMode(file.content, "merge");
-	} else if (providerIds.length > 0) {
-		providerPatchContent = withOpenClawProviderMode(file.content, "replace");
-	}
-	return {
-		apply: true,
-		content: removeNativeOpenClawCredentials(
-			mergeProviderDeletes("openclaw", providerPatchContent, deletedProviderIds),
-			previousNativeProviderIds.filter((id) => !nativeIds.has(id) && !projectedIds.has(id)),
-		),
-		providerIds,
-		nativeCredentialProviderIds: [...nativeIds],
-	};
-}
-
-function removeNativeOpenClawCredentials(content: string, providerIds: readonly string[]): string {
-	if (providerIds.length === 0) return content;
-	const patch = recordValue(JSON.parse(content) as unknown);
-	if (!patch) throw new Error("OpenClaw provider patch must be an object");
-	const models = recordValue(patch.models) ?? {};
-	const providers = { ...(recordValue(models.providers) ?? {}) };
-	for (const id of providerIds) providers[id] = { apiKey: null, auth: null, baseUrl: null };
-	models.providers = providers;
-	patch.models = models;
-	return `${JSON.stringify(patch, null, 2)}\n`;
-}
-type ProviderPatchRuntime = "hermes" | "openclaw";
-function providerPatchRoot(
-	runtime: ProviderPatchRuntime,
-	content: string,
-): Record<string, unknown> | null {
-	if (runtime === "hermes" && !content.trim()) return null;
-	return recordValue(runtime === "openclaw" ? JSON.parse(content) : parseYaml(content));
-}
-function providerPatchProviders(
-	runtime: ProviderPatchRuntime,
-	root: Record<string, unknown>,
-): Record<string, unknown> | null {
-	const container = runtime === "openclaw" ? recordValue(root.models) : root;
-	return container ? recordValue(container.providers) : null;
-}
-function providerIdsFromPatch(runtime: ProviderPatchRuntime, content: string): Set<string> {
-	const root = providerPatchRoot(runtime, content);
-	const providers = root ? providerPatchProviders(runtime, root) : null;
-	if (!providers) return new Set();
-	return new Set(
-		Object.entries(providers)
-			.filter(([, value]) => value !== null)
-			.map(([providerId]) => providerId),
-	);
-}
-function withOpenClawProviderMode(patchContent: string, mode: "merge" | "replace"): string {
-	const parsed = JSON.parse(patchContent) as unknown;
-	const root = recordValue(parsed);
-	if (!root) return patchContent;
-	const patch = { ...root };
-	const models = { ...(recordValue(patch.models) ?? {}), mode };
-	patch.models = models;
-	return `${JSON.stringify(patch, null, 2)}\n`;
-}
-
-const HERMES_DIRECT_MODEL_FIELDS = [
-	"base_url",
-	"api_key",
-	"api",
-	"key_env",
-	"api_mode",
-	"auth_mode",
-] as const;
-const HERMES_GENERATED_PROVIDER_FIELDS = [
-	"name",
-	"api",
-	"url",
-	"base_url",
-	"default_model",
-	"model",
-	"models",
-	"discover_models",
-	"transport",
-	"api_mode",
-	"key_env",
-	"api_key",
-	"type",
-	"auth_type",
-] as const;
-function applyHermesProviderConfig(
-	context: HermesConfigTransaction,
-	patch: Record<string, unknown>,
-	deletedProviderIds: readonly string[],
-	preserveModelSelection = false,
-): void {
-	const patchModel = recordValue(patch.model) ?? {};
-	const modelKeys = new Set<string>([...HERMES_DIRECT_MODEL_FIELDS, ...Object.keys(patchModel)]);
-	for (const key of preserveModelSelection ? [] : [...modelKeys].sort()) {
-		const value = Object.hasOwn(patchModel, key) ? patchModel[key] : undefined;
-		reconcileHermesConfigValue(context, `model.${key}`, value === null ? undefined : value);
-	}
-	if (
-		!preserveModelSelection &&
-		!Object.hasOwn(patchModel, "provider") &&
-		deletedProviderIds.length > 0
-	) {
-		const currentProvider = getHermesResolvedConfigValue(context.context, "model.provider");
-		if (currentProvider.exists && typeof currentProvider.value !== "string") {
-			throw new Error("Hermes config field model.provider must be a string");
-		}
-		const managedSelectors = new Set(
-			deletedProviderIds.flatMap((providerId) => [providerId, `custom:${providerId}`]),
-		);
-		if (currentProvider.exists && managedSelectors.has(currentProvider.value as string)) {
-			reconcileHermesConfigValue(context, "model.provider", undefined);
-			if (!Object.hasOwn(patchModel, "default")) {
-				reconcileHermesConfigValue(context, "model.default", undefined);
-			}
-		}
-	}
-
-	const currentValue = getHermesRawConfigValue(context, "providers");
-	if (currentValue.exists && !isPlainRecord(currentValue.value)) {
-		throw new Error("Hermes config field providers must be an object");
-	}
-	const currentProviders: Record<string, unknown> =
-		currentValue.exists && isPlainRecord(currentValue.value) ? currentValue.value : {};
-	for (const [providerId, provider] of Object.entries(currentProviders)) {
-		if (provider !== undefined && provider !== null && !isPlainRecord(provider)) {
-			throw new Error(`Hermes provider ${providerId} must be an object`);
-		}
-	}
-	const nextProviders: Record<string, unknown> = { ...currentProviders };
-	for (const providerId of deletedProviderIds) delete nextProviders[providerId];
-
-	const patchProviders = recordValue(patch.providers) ?? {};
-	for (const [providerId, providerPatch] of Object.entries(patchProviders)) {
-		if (providerPatch === null) {
-			delete nextProviders[providerId];
-			continue;
-		}
-		if (!isPlainRecord(providerPatch)) continue;
-		const existingProvider = nextProviders[providerId];
-		if (
-			existingProvider !== undefined &&
-			existingProvider !== null &&
-			!isPlainRecord(existingProvider)
-		) {
-			throw new Error(`Hermes provider ${providerId} must be an object`);
-		}
-		const nextProvider: Record<string, unknown> = isPlainRecord(existingProvider)
-			? { ...existingProvider }
-			: {};
-		for (const key of HERMES_GENERATED_PROVIDER_FIELDS) delete nextProvider[key];
-		let wroteGeneratedField = false;
-		for (const [key, value] of Object.entries(providerPatch)) {
-			if (value === null) {
-				delete nextProvider[key];
-				continue;
-			}
-			nextProvider[key] = value;
-			wroteGeneratedField = true;
-		}
-		const hasUserOwnedField = Object.keys(nextProvider).some(
-			(key) => !(HERMES_GENERATED_PROVIDER_FIELDS as readonly string[]).includes(key),
-		);
-		if (wroteGeneratedField || hasUserOwnedField) nextProviders[providerId] = nextProvider;
-		else delete nextProviders[providerId];
-	}
-
-	if (Object.keys(nextProviders).length === 0 && Object.keys(currentProviders).length === 0) return;
-	reconcileHermesConfigValue(
-		context,
-		"providers",
-		Object.keys(nextProviders).length > 0 ? nextProviders : undefined,
-	);
-}
-function mergeProviderDeletes(
-	runtime: ProviderPatchRuntime,
-	patchContent: string,
-	deletedProviderIds: readonly string[],
-): string {
-	if (deletedProviderIds.length === 0) return patchContent;
-	const root = providerPatchRoot(runtime, patchContent);
-	if (!root) return patchContent;
-	const patch = { ...root };
-	const container =
-		runtime === "openclaw" ? { ...(recordValue(patch.models) ?? { mode: "merge" }) } : patch;
-	const existingProviders = recordValue(container.providers);
-	const providers = existingProviders ? { ...existingProviders } : {};
-	for (const providerId of deletedProviderIds) {
-		providers[providerId] = null;
-	}
-	container.providers = providers;
-	if (runtime === "openclaw") {
-		patch.models = container;
-		if (deletedProviderIds.some(isClawdiManagedProviderId)) {
-			const agents = { ...(recordValue(patch.agents) ?? {}) };
-			const defaults = { ...(recordValue(agents.defaults) ?? {}) };
-			defaults.memorySearch = null;
-			agents.defaults = defaults;
-			patch.agents = agents;
-			const memory = { ...(recordValue(patch.memory) ?? {}) };
-			const search = { ...(recordValue(memory.search) ?? {}) };
-			if (!Object.hasOwn(search, "provider")) search.provider = null;
-			if (!Object.hasOwn(search, "model")) search.model = null;
-			memory.search = search;
-			patch.memory = memory;
-		}
-	}
-	return runtime === "openclaw"
-		? `${JSON.stringify(patch, null, 2)}\n`
-		: `${stringifyYaml(patch).trimEnd()}\n`;
-}
-function staleProviderIds(
-	previousProviderIds: Set<string>,
-	activeProviderIds: Set<string>,
-): string[] {
-	return [...previousProviderIds]
-		.filter((providerId) => !activeProviderIds.has(providerId))
-		.sort((left, right) => left.localeCompare(right));
-}
-export function openClawGatewayHostedPatch(
-	manifest: RuntimeManifest,
-	secretValues: Record<string, string> | undefined,
-	ownerBrowserBootstrapSupported: boolean,
-): Record<string, unknown> | null {
-	const allowedOrigins = openClawControlUiAllowedOrigins(manifest);
-	const trustedProxies = openClawGatewayTrustedProxies(manifest);
-	const gatewayToken = manifest.openclawGatewayAuth
-		? runtimeSecretValue(secretValues ?? {}, manifest.openclawGatewayAuth.tokenRef)
-		: null;
-	const nativeAuth = manifest.openclawGatewayAuth;
-	if (nativeAuth?.activation.enabled !== true) {
-		throw new Error("OpenClaw native auth capability is unavailable");
-	}
-	if (manifest.openclawGatewayAuth && !gatewayToken) {
-		throw new Error("OpenClaw native gateway token is unavailable");
-	}
-	if (allowedOrigins.length === 0 && !gatewayToken && !manifest.locale) return null;
-	return {
-		...(manifest.locale
-			? {
-					agents: {
-						defaults: {
-							userTimezone: manifest.locale.timezone,
-						},
-					},
-				}
-			: {}),
-		gateway: {
-			mode: "local",
-			...(trustedProxies.length > 0 ? { trustedProxies } : {}),
-			...(gatewayToken || allowedOrigins.length > 0
-				? {
-						...(nativeAuth ? { port: 18789, bind: "lan" } : {}),
-						...(gatewayToken
-							? {
-									auth: {
-										mode: "token",
-										token: gatewayToken,
-									},
-								}
-							: {}),
-						...(allowedOrigins.length > 0
-							? {
-									controlUi: {
-										allowedOrigins,
-										...(nativeAuth
-											? {
-													basePath: openClawControlUiBasePath(manifest),
-													dangerouslyAllowHostHeaderOriginFallback: false,
-													dangerouslyDisableDeviceAuth: ownerBrowserBootstrapSupported
-														? null
-														: true,
-												}
-											: {}),
-									},
-								}
-							: {}),
-					}
-				: {}),
-		},
-	};
-}
-function jsonMergePatchIsApplied(current: unknown, patch: unknown): boolean {
-	if (!isPlainRecord(patch)) return canonicalJsonEqual(current, patch);
-	if (!isPlainRecord(current)) {
-		if (current !== undefined) return false;
-		// OpenClaw canonicalizes deletion-only patches by omitting their empty parents.
-		return Object.values(patch).every(
-			(value) =>
-				value === undefined ||
-				value === null ||
-				(isPlainRecord(value) && jsonMergePatchIsApplied(undefined, value)),
-		);
-	}
-	return Object.entries(patch).every(([key, value]) =>
-		value === undefined
-			? true
-			: value === null
-				? !Object.hasOwn(current, key)
-				: jsonMergePatchIsApplied(current[key], value),
-	);
-}
-export function openClawConfigPatchIsApplied(
-	context: OpenClawHostedContext,
-	patch: Record<string, unknown>,
-): boolean {
-	try {
-		const current = JSON.parse(readFileSync(context.configPath, "utf-8")) as unknown;
-		return jsonMergePatchIsApplied(current, patch);
-	} catch {
-		return false;
-	}
-}
-function openClawControlUiBasePath(manifest: RuntimeManifest): string {
-	const system = manifest.projection?.system;
-	if (!isPlainRecord(system)) return "/";
-	const value = system.openclawControlUiBasePath;
-	if (typeof value !== "string" || !value.startsWith("/")) return "/";
-	return value === "/" ? "/" : value.replace(/\/$/, "");
-}
-function applyOpenClawGatewayHostedProjection(
-	command: string,
-	manifest: RuntimeManifest,
-	secretValues: Record<string, string> | undefined,
-	context: OpenClawHostedContext,
-	workspaceRoot: string,
-	ownerBrowserBootstrapSupported: boolean,
-): void {
-	const patch = openClawGatewayHostedPatch(manifest, secretValues, ownerBrowserBootstrapSupported);
-	if (!patch || openClawConfigPatchIsApplied(context, patch)) return;
-	runRuntimeUserCommand(
-		command,
-		["config", "patch", "--stdin"],
-		`${JSON.stringify(patch, null, 2)}\n`,
-		context.home,
-		workspaceRoot,
-	);
-}
-function openClawControlUiAllowedOrigins(manifest: RuntimeManifest): string[] {
-	const system = manifest.projection?.system;
-	if (!isPlainRecord(system)) return [];
-	const raw = system.openclawControlUiAllowedOrigins;
-	if (!Array.isArray(raw)) return [];
-	const seen = new Set<string>();
-	const origins: string[] = [];
-	for (const value of raw) {
-		if (typeof value !== "string") continue;
-		const origin = value.trim();
-		if (!origin || seen.has(origin)) continue;
-		seen.add(origin);
-		origins.push(origin);
-	}
-	return origins;
-}
-function openClawGatewayTrustedProxies(manifest: RuntimeManifest): string[] {
-	const system = manifest.projection?.system;
-	if (!isPlainRecord(system)) return [];
-	const raw = system.openclawGatewayTrustedProxies;
-	return Array.isArray(raw)
-		? raw.filter((value): value is string => typeof value === "string")
-		: [];
 }
